refactor: folder name from datasets to datamodules and readme fix
