+++ personalized_active_learning/datamodules/definitions/__init__.py
@@ -1,0 +1,3 @@
+"""Definitions of concrete datamodules implementations."""
+
+from .unhealthy import UnhealthyDataModule
