--- personalized_active_learning/datasets/definitions/__init__.py
@@ -1,3 +1,0 @@
-"""Definitions of concrete datasets implementations."""
-
-from .unhealthy import UnhealthyDataModule
