+++ personalized_active_learning/learning/training.py
@@ -30,7 +30,7 @@ def train_test(
         model (IModel):  A model defined in `personalized_active_learning.models`
                 directory. Used for a main training. (Imorted definition, eg. `Baseline`).
         datamodule (_type_): A dataset which derives from the
-                `personalized_active_learning.datasets.base` class.
+                `personalized_active_learning.datamodules.base` class.
         logger (WandbLogger): Instantion of prepared `WandbLogger`.
         epochs (int, optional): Number of training epochs. Defaults to 6.
         lr (float, optional): Learning rate for a model. Defaults to 1e-2.
